Build NSS into the docker image
This resulted in a somewhat flabby image, but I think that it should
make builds a lot faster.  At least it isn't the 2.4G that my first
attempt produced.  I managed to trim it down a fair bit.

If we want to commit to a single NSS version, this could be trimmed down
a lot more.  Right now, this assumes that you want to pull the latest
NSS and NSPR and build them.  To make that process faster, I keep the
NSS and NSPR checkouts and the object files from the build.  That should
make most builds a no-op.

I learned something interesting about mercurial here.  If you remove a
directory, it will happily update even if there are changes to that
directory.  That allowed me to remove some of the larger directories
from NSS.

As a bonus, I've disabled building of tests for NSS, which drives the
compilation time down a lot.  That benefit should be noticeable in other
ways.